feat: Pull out polyfill and pfelement updates
